# [Inductor] default block size for head_dim = 256 for flex attention